+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,15 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - Unreleased
+
+### Added
+
+- The `portable` shell option, which restricts the shell to portable constructs
+  so that scripts can be checked for portability. When it is set, the shell now
+  rejects the non-portable `;;&` and `;|` case terminators. More checks will be
+  added in future releases.
+
 ## [3.2.1] - 2026-06-21
 
 ### Changed